fix gcc issues

+++ FIRE/src/Draw.h
@@ -1,6 +1,7 @@
 #ifndef FIRE_DRAW_H
 #define FIRE_DRAW_H
 #include <FIRE/MeshType.h>
+#include <cstdint>
 namespace FIRE
 {
 class VertexLayout;
@@ -9,8 +10,8 @@ class Draw
 public:
     virtual ~Draw() = default;
     virtual void Clear() = 0;
-    virtual void DoDraw(VertexLayout& layout, MeshPrimitives primitives, size_t count) = 0;
-    virtual void DoDrawIndexed(VertexLayout& layout, MeshPrimitives primitives, size_t count) = 0;
+    virtual void DoDraw(VertexLayout& layout, MeshPrimitives primitives, uint32_t count) = 0;
+    virtual void DoDrawIndexed(VertexLayout& layout, MeshPrimitives primitives, uint32_t count) = 0;
     virtual void ToggleWireframe() = 0;
 };
 } // namespace FIRE
